feat: unify follow button
